Finish up the particle engine
