+++ shared/randomPoint/randomPointInSphere.ts
@@ -14,3 +14,27 @@ export function randomPointInSphere(radius = 1) {
 	const z = r * cosPhi;
 	return [x, y, z] as const;
 }
+
+export function randomPointOnSphere(radius = 1) {
+	const [x, y, z] = [
+		generateGaussian(),
+		generateGaussian(),
+		generateGaussian(),
+	];
+	const normalized = Math.sqrt(x * x + y * y + z * z);
+	return [
+		(x / normalized) * radius,
+		(y / normalized) * radius,
+		(z / normalized) * radius,
+	] as const;
+}
+
+function generateGaussian() {
+	let u = 0;
+	let v = 0;
+	// Converting [0,1) to (0,1) to avoid log(0) error
+	while (u === 0) u = Math.random();
+	while (v === 0) v = Math.random();
+
+	return Math.sqrt(-2.0 * Math.log(u)) * Math.cos(2.0 * Math.PI * v);
+}
